1. разделение кода на отдельные файлы. 2. минимальные правки после комментариев Алексея. 3. добавлено API wiki, но взаимодействия с ботом пока что нет.
