+++ wiki/wikiAPI.go
@@ -1,0 +1,71 @@
+package wiki
+
+import (
+	"encoding/json"
+	"io"
+	"log"
+	"net/http"
+)
+
+type SearchResults struct {
+	ready   bool
+	Query   string
+	Results []Result
+}
+
+type Result struct {
+	Name, Description, URL string
+}
+
+/*func (sr *SearchResults) UnmarshalJSON(bs []byte) error {
+	array := []interface{}{}
+	if err := json.Unmarshal(bs, &array); err != nil {
+		return err
+	}
+	sr.Query = array[0].(string)
+	for i := range array[1].([]interface{}) {
+		sr.Results = append(sr.Results, Result{
+			array[1].([]interface{})[i].(string),
+			array[2].([]interface{})[i].(string),
+			array[3].([]interface{})[i].(string),
+		})
+	}
+	return nil
+}*/
+
+func WikipediaAPI(request string) (answer []string) {
+
+	//Создаем срез на 3 элемента
+	s := make([]string, 3)
+
+	//Отправляем запрос
+	if response, err := http.Get(request); err != nil {
+		s[0] = "Wikipedia is not respond"
+	} else {
+		defer response.Body.Close()
+
+		//Считываем ответ
+		contents, err := io.ReadAll(response.Body)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		//Отправляем данные в структуру
+		sr := &SearchResults{}
+		if err = json.Unmarshal([]byte(contents), sr); err != nil {
+			s[0] = "Something going wrong, try to change your question"
+		}
+
+		//Проверяем не пустая ли наша структура
+		if !sr.ready {
+			s[0] = "Something going wrong, try to change your question"
+		}
+
+		//Проходим через нашу структуру и отправляем данные в срез с ответом
+		for i := range sr.Results {
+			s[i] = sr.Results[i].URL
+		}
+	}
+
+	return s
+}
